+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ------
 
 **ENHANCEMENTS**
+- Allow configuration of static and dynamic node priorities in compute resources via the ParallelCluster configuration YAML file.
 
 **CHANGES**